+++ tests/testthat/test-sampling.R
@@ -4,12 +4,10 @@ test_that("voxel sampling works",
   o = tempfile(fileext = ".las")
   vox = sampling_voxel(5)
   write = write_las(o)
-  pipeline = vox + write
-  exec(pipeline, on = f)
+  pipeline = vox + write + summarise()
+  ans = exec(pipeline, on = f)
 
-  las = read_las(o, expose = "xyzi")
-
-  expect_equal(dim(las), c(7864L, 4L))
+  expect_equal(ans$summary$npoints, 7864L)
 })
 
 test_that("pixel sampling works",
@@ -18,10 +16,20 @@ test_that("pixel sampling works",
    o = tempfile(fileext = ".las")
    vox = sampling_pixel(5)
    write = write_las(o)
-   pipeline = vox + write
-   exec(pipeline, on = f)
+   pipeline = vox + write + summarise()
+   ans = exec(pipeline, on = f)
 
-   las = read_las(o, expose = "xyzi")
+   expect_equal(ans$summary$npoints, 3042L)
+})
 
-   expect_equal(dim(las), c(3042L, 4L))
+test_that("poisson sampling works",
+{
+  f = system.file("extdata", "Topography.las", package="lasR")
+  o = tempfile(fileext = ".las")
+  vox = sampling_poisson(5)
+  write = write_las(o)
+  pipeline = vox + write + summarise()
+  ans = exec(pipeline, on = f)
+
+  expect_equal(ans$summary$npoints, 3742L)
 })